+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.0 (2022-09-04)
 - Added the `set_timeout_async!` macro. It consumes an identifier that is an async function.
   Async functions are currently not supported as this required an unstable async closure feature.
+- small fix in `set_interval!`
 
 # 1.1.1 (2022-05-02)
 - small internal improvements